requirements.txt
psycopg2-binary
python-dotenv
fastapi
uvicorn
requests
tqdm
paramiko
